fix: update, delete 시 500 에러 오류 수정

+++ src/main/java/com/example/calendar/service/WriterServiceImpl.java
@@ -12,10 +12,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -31,7 +28,8 @@ public class WriterServiceImpl implements WriterService{
 
     @Override
     public void editWriterInfo(WriterUpdateInput updateInput) {
-        writerRepository.update(updateInput);
+        Writer writer = getWriterById(updateInput.getId());
+        if(writer != null) writerRepository.update(updateInput);
     }
 
     @Override
@@ -55,7 +53,8 @@ public class WriterServiceImpl implements WriterService{
 
     @Override
     public void withdraw(WriterDeleteInput deleteInput) {
-        writerRepository.delete(deleteInput);
+        Writer writer = getWriterById(deleteInput.getId());
+        if(writer != null) writerRepository.delete(deleteInput);
     }
 
     /**
